Move storage classes into 'server' package.
Made value() methods for Command and Event public. They are likely to queried from everywhere. One of the cases was from the 'server' package already.

The main reason of moving AggregateRoot into the 'server' was that the dispatch() method isn't really a part of the public API or a root. It's called by AbstractRepository only. Having AggregateRoot class inside the 'server' package slightly obscures the design, but it's a small price as compared with the confusion of having unused public method.