Log warnings instead of throwing exceptions when wrong method modifier encountered.
It's not that kind of error, which is fatal to the app. It's just a programmer's mistake, which doesn't affect the framework or the customer's app. The only negative side of having wrong modifier in the user's app is accidental call of the method from another part of the code.