Remove iterating command router
Its naīve implementation won't stand in a distributed environment.
Anyway, a more complex routing scenario would need a PM state, and,
most likely conditional logic, that handles rejections or errors
occurred during processing of already posted commands. So, it's not
going to be a sequence in a simple way that it was implemented in the
iterating command router.