Add a method for soft repository lookup and use it for `Mirror` entities
When aggregate querying is disabled, there is no repository for
`Mirror` entities in the system at all. Thus, using the
`findRepository(Mirror.class)` method leads to immediate
`IllegalStateException`.

To fix this, a new implementation of `findRepository(Class)` is
added, now it returns an `Optional.empty()` in case the repository for
entity state type is not found. An old method is renamed to
`repositoryFor(Class)`.