Make ID parameter first
This order is more natural because we an ID is a "pointer" to an entity. Also, we're going to have more assertion methods that would accept an instance of an entity state. Creation of an entity state (which is a Protobuf message) is a big block of code. It would be more readable if the ID parameter comes before such a block. So, to make the order consistent for all assertion methods, we're making ID coming first for all.