Bring back the `[(required) = false]` workaround for PM tests
Removing this workaround leads to heavily verbose setup for some of the test cases, to the point where it makes it harder to understand the original meaning of the test.

To address the problem, the concept of having `repository.create(ID)` which returns an invalid entity, which then cannot be loaded, should be revisited as a whole.